Removed monorepo configuration
The main benefit of the monorepo setup was subdomains. However, I'm not so sure that is valuable at this point. For one, that would require separate projects for each app, meaning separate analytics and other components. Secondly, app router already supports folder level colocation and shared layouts, which is what I was trying to do but in a more complex way with monorepos.